implement provisory block on account deletion

+++ backend/src/routes/controllers/account/deleteAccount.controller.ts
@@ -2,6 +2,7 @@ import { NextFunction, Request, Response } from "express";
 import UserRepository from "../../../repositories/User.repository";
 import HttpError from "../../../utils/HttpError";
 import HTTPCodes from "simple-http-codes";
+import OrdersRepository from "../../../repositories/Orders.repository";
 
 export default async function deleteAccountController(
 	_req: Request,
@@ -9,6 +10,30 @@ export default async function deleteAccountController(
 	next: NextFunction
 ) {
 	try {
+		//Check if its possible to delete account
+		const orders_response = await OrdersRepository.has({
+			email: res.locals.email,
+			delivered: true,
+		});
+
+		if (orders_response.error) {
+			throw new HttpError(
+				HTTPCodes.ServerError.BAD_GATEWAY,
+				orders_response.message,
+				{
+					error: orders_response.error,
+					context: "checking if account has ongoing orders",
+				}
+			);
+		}
+
+		if (orders_response.data) {
+			return res
+				.status(HTTPCodes.ClientError.CONFLICT)
+				.json({ message: "Can´t delete account with ongoing orders" });
+		}
+
+		//delete account
 		const db_response = await UserRepository.delete({
 			email: res.locals.email,
 		});
